Assignment 5.
First, I created an Aggregate variable for the 12 SF-12 variables which became the Overall Health Variable. Then I made a data list of the first 50 to check that it worked. Then, using summary statistics, I found the mean, mode, std dev, first qrt, and third qrt of the Aggregate. Then, I made a histogram of the variable. After that, I created a categorical variable for both Married and Education year. The Marriage one was either [1] Married or [2] Not Married. The year of education I did less than 12 yrs (High School Diploma)=1, 12 years to Less than 14 is Highschool Diploma and everything above that was College degree.